refactor(memory): unify capture review under memory
Moves raw capture review into a shared Memory component so
/memory/captures is primary while /archive stays compatible.

Updates cockpit, dashboard, and navigation links to point at Memory
review while tests cover both base paths.
